refactor(list): use cell and row current variant for loading placeholder
