+++ src/components/List/variants/index.ts
@@ -1,0 +1,11 @@
+import * as variantExplorer from './variantExplorer'
+import * as variantProduct from './variantProduct'
+import * as variantTable from './variantTable'
+
+export const variants = {
+  explorer: variantExplorer,
+  product: variantProduct,
+  table: variantTable,
+}
+
+export type ListVariant = keyof typeof variants
