+++ src/components/List/variants/variantTable.tsx
@@ -2,14 +2,14 @@ import { Theme, css } from '@emotion/react'
 import styled from '@emotion/styled'
 import PropTypes from 'prop-types'
 import { KeyboardEvent, MouseEvent, useCallback } from 'react'
-import * as animations from '../../utils/animations'
-import Box from '../Box'
-import Checkbox from '../Checkbox'
-import Tooltip from '../Tooltip'
-import BaseCell from './Cell'
-import SortIcon from './SortIcon'
-import { useListContext } from './context'
-import { ListRowProps } from './types'
+import * as animations from '../../../utils/animations'
+import Box from '../../Box'
+import Checkbox from '../../Checkbox'
+import Tooltip from '../../Tooltip'
+import BaseCell from '../Cell'
+import SortIcon from '../SortIcon'
+import { useListContext } from '../context'
+import { ListRowProps } from '../types'
 
 export const Cell = styled(BaseCell)``
